refactor(audio-slider): move to specific folder
